requirements.txt
torch>=2.2
torchaudio>=2.2
torchmetrics>=1.4
numpy>=1.24
tqdm>=4.66
hydra-core>=1.3
omegaconf>=2.3
comet-ml>=3.44
python-dotenv>=1.0
soundfile>=0.12
matplotlib>=3.8
pandas>=2.0
pystoi>=0.4
librosa>=0.10
